+++ src/core/dense_match/DisparityValidator.h
@@ -26,7 +26,7 @@ public:
                                const cv::Mat &right) const;
 
 private:
-    DenseMatchConfig m_cfg;
+    DenseMatchConfig _config;
 };
 
 } // namespace xjw::dense_match